Stop recommending removal of ClamAV when ImunifyAV is installed.
Case CPANEL-34196: When ImunifyAV (free version) was added to
the product, the Security Advisor was updated to check whether
ClamAV was also installed, and if so, recommend removing it on
the grounds that it would be redundant. ClamAV is not redundant,
however, because it is still required for a setting in WHM that
enables scanning messages that pass through Exim.